refactor: migrate to official MCP SDK and centralize FastMCP instance initialization in __init__.py
